src/EHR_service/requirements.txt
Django==5.0.2
djangorestframework==3.14.0
django-cors-headers==4.3.1
python-dotenv==1.0.1
djongo==1.3.6
pymongo==4.6.1
PyJWT==2.8.0
django-filter==24.1
celery==5.3.6
redis==5.0.1
drf-yasg==1.21.7
pytest==8.0.2
pytest-django==4.8.0
black==24.2.0
flake8==7.0.0
isort==5.13.2
sentry-sdk==1.40.4
python-json-logger==2.0.7
requests==2.31.0 